DeskHub/requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=3.0
pyodbc>=4.0
SQLAlchemy>=1.4
python-dotenv>=1.0
